+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `ModelArrows`
 - `ModelText`
 - `Solid.Intersects(Plane p, out List<Polygon> result)`
+- `Vector3.IsUnitized()`
 
 ### Changed